+++ httpclient-vertx-5/src/main/java/io/fabric8/kubernetes/client/vertx/VertxHttpClientBuilder.java
@@ -180,6 +180,12 @@ public class VertxHttpClientBuilder<F extends HttpClient.Factory>
 
     final WebSocketClientOptions wsOptions = createWebSocketClientOptions(protocols);
 
+    // Since the factory is not explicitly typed as VertxHttpClientFactory, we need to check the type before invoking
+    // additionalConfig as to not break other implementations of HttpClient.Factory for VertxHttpClientBuilder.
+    if (clientFactory instanceof VertxHttpClientFactory) {
+      ((VertxHttpClientFactory) clientFactory).additionalConfig(wsOptions, httpOptions, poolOptions);
+    }
+
     return VertxHttpClient.createWithWebSocketOptions(
         this,
         new AtomicBoolean(),